backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
python-dotenv
aiokafka
